+++ src/com/euhedral/game/Entities/Enemy/Enemy.java
@@ -371,7 +371,7 @@ public class Enemy extends MobileEntity {
     }
 
     @Override
-    public void resurrect(int x, int y) {
+    public void resurrect(double x, double y) {
         commonInit();
         explosion.playedOnce = false;
         super.resurrect(x, y);
